Make async hooks more modular
